requirements.txt
requests>=2.31
python-whois>=0.9
rich>=13.0
python-dotenv>=1.0
